requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic-settings==2.5.2
Pillow==10.4.0
itsdangerous==2.2.0
requests==2.32.3
python-dotenv==1.0.1
numpy==2.1.1
lxml==5.3.0
